faker/requirements.txt
pydantic==2.11.4
Faker==37.1.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
